flamethrower mechanics

+++ Assets/Scripts/PlayerHealth.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class PlayerHealth : MonoBehaviour, IHealthSystem
 
 {
-    public int Health { get; set; }
-    public int CurrentHealth { get; set; }
+    public float Health { get; set; }
+    public float CurrentHealth { get; set; }
     [SerializeField] private HealthBar HealthBar;
 
     // Start is called before the first frame update
@@ -23,7 +23,7 @@ public class PlayerHealth : MonoBehaviour, IHealthSystem
     {
     }
 
-    public void Damage(int damage)
+    public void Damage(float damage)
     {
         CurrentHealth -= damage;
         if (CurrentHealth < 0) CurrentHealth = 0;
@@ -32,7 +32,7 @@ public class PlayerHealth : MonoBehaviour, IHealthSystem
 
     public void Heal()
     {
-        int newHealth = CurrentHealth;
+        float newHealth = CurrentHealth;
         if (newHealth == CurrentHealth)
         {
             CurrentHealth += 10;
